# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAPhoto API
  description: |-
    This is the API for WASAPhoto homework 1 by 1894389
  version: 0.1.0
...
paths:
...
  /users/{id}/followers/{follower_id}:
    parameters:
        - $ref: "#/components/parameters/follower_id"
        - $ref: '#/components/parameters/identifier'
        
    put:
      tags: ["followers"]
      summary: Follows a specific user
      description: Allows a user to follow another user. 
      operationId: followUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
          
    delete:
      tags: ["followers"]
      summary: Unfollows a specific user
      description: Unfollow another one,removing him from your list of followed users.
      operationId: unfollowUser
      
      responses:
        '204':
          $ref: "#/components/responses/no_content"
        '401':
          $ref: "#/components/responses/unauthorized"
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"      
          
      security:
        - bearerAuth: []
# --------------------------------------------------------------
...
  /users/{id}/photos/{photo_id}/comments/{comment_id}:
    parameters: 
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/photo_id'
        - $ref: '#/components/parameters/comment_id'
        
    delete:
      tags: ["comments"]
      summary: Remove a comment from a photo
      description: Allows a user to remove a comment from a photo he posted.
      operationId: uncommentPhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: []
# --------------------------------------------------------------
  /users/{id}/photos/{photo_id}/likes/{like_id}:
    parameters: 
        - $ref: '#/components/parameters/identifier'
        - $ref: '#/components/parameters/photo_id'
        - $ref: '#/components/parameters/like_id'
  
    put:
      tags: ["likes"]
      summary: Add a like to a photo
      description: Adds a like to a user's photo. A user can only like a photo once.
      operationId: likePhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
      
      security:
        - bearerAuth: [] 
    
    delete:
      tags: ["likes"]
      summary: Remove a like from a photo
      description: Removes a like from a user's photo
      operationId: unlikePhoto
      
      responses:
        '204':
          $ref: '#/components/responses/no_content'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: "#/components/responses/not_found"
        '500':
          $ref: "#/components/responses/internal_server_error"
          
      security:
        - bearerAuth: [] 
      
# --------------------------------------------------------------
components:
  parameters:
    identifier:
      name: id
      in: path
      description: It's a user's unique id
      required: true
      schema:
        description: String between 3-16 characters
        type: string
        pattern: '^.*?$'
        example: Micheal0Jackson
        minLength: 3
        maxLength: 16
        readOnly: true
...
    follower_id:
      name: follower_id
      in: path
      description: A follower's identifier.
      required: true
      schema: 
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "Pikachu1"
    photo_id:
      name: photo_id
      in: path
      description: A photo identifier
      required: true
      schema: 
        description: Int64 between 0-9999
        type: integer
        format: int64
        minimum: 0
        maximum: 9999
        example : 999
        readOnly: true
    comment_id:
      name: comment_id
      in: path
      description: A comment identifier
      required: true
      schema:
        $ref: "#/components/schemas/comment_identifier/properties/comment_id"
      example: 123
    like_id:
      name: like_id
      in: path
      description: A like unique identifier.
      required: true
      schema: 
        $ref: "#/components/schemas/user_nickname/properties/nickname"
      example: "Mami"
      
# --------------------------------------------------------------
...
  responses:
...
    no_content:
      description: This response is associated to the 204 http status (There is no content to send for this request)
...
    unauthorized:
      description: This response is associated to the 401 http status (The client must authenticate itself to get the requested response)
    forbidden:
      description: This response is associated to the 403 http status (The client does not have access rights to the content)
    not_found:
      description: This response is associated to the 404 http status (The server cannot find the requested resource.)
    internal_server_error:
      description: This response is associated to the 500 http status (The server has encountered a situation it does not know how to handle.)

# --------------------------------------------------------------
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT